fix: complete route group rename from (dashboard) to dashboard-v2
- Renamed app/(dashboard) to app/dashboard-v2 to avoid Next.js 15 parentheses bug
- Updated middleware.ts to handle /dashboard-v2 path with array matcher
- Fixed test imports to use dashboard-v2 instead of (dashboard)
- Moved __tests__/app/(dashboard) to __tests__/app/dashboard-v2
- Fixed User type lint errors by adding missing status property in tests

Build successful (19 pages), all tests passing (357/357)
